refactor(structure): move annotations to dedicated folder
