+++ .eslintrc
@@ -27,7 +27,7 @@
     "@typescript-eslint/no-explicit-any": "error",
     "import/prefer-default-export": "off",
     "import/no-internal-modules": ["error", {
-      "allow": ["src/*"]
+      "allow": ["annotations/*"]
     }],
     "import/order": [
       "error",
