+++ index.ts
@@ -1,9 +1,11 @@
+/* eslint-disable import/no-internal-modules */
+
 import "reflect-metadata";
 
-export * from "./src/autowire";
-export * from "./src/debugger";
-export * from "./src/env";
-export * from "./src/injectable";
+export * from "./src/annotations/autowire";
+export * from "./src/annotations/debugger";
+export * from "./src/annotations/env";
+export * from "./src/annotations/injectable";
 export * from "./src/injection-context";
 export * from "./src/throwable";
 export * from "./src/utils";
